backend/requirements.txt
# ===============================================
# VITEVITEAPP - BACKEND PRODUCTION REQUIREMENTS
# Python 3.11+
# ===============================================

# ========== CORE FRAMEWORK ==========
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
email-validator==2.1.0

# ========== DATABASE ==========
sqlalchemy==2.0.27
asyncpg==0.29.0          # PostgreSQL async driver
alembic==1.13.1          # Migrations
psycopg2-binary==2.9.9   # PostgreSQL sync (pour Alembic)

# ========== VALIDATION ==========
pydantic==2.6.1
pydantic-settings==2.1.0

# ========== AUTHENTICATION ==========
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
bcrypt==4.1.2

# ========== AI & ML ==========
google-generativeai==0.3.2    # Gemini
openai==1.12.0                # Fallback
scikit-learn==1.4.0
xgboost==2.0.3
numpy==1.26.3
pandas==2.2.0

# ========== VOICE (OPTIONAL) ==========
elevenlabs==0.2.27
SpeechRecognition==3.10.1

# ========== CACHE & QUEUE ==========
redis==5.0.1
aioredis==2.0.1
celery==5.3.6            # Background tasks

# ========== HTTP & REQUESTS ==========
httpx==0.26.0
aiofiles==23.2.1
requests==2.31.0

# ========== MONITORING & LOGGING ==========
structlog==24.1.0
sentry-sdk[fastapi]==1.40.0

# ========== TESTING ==========
pytest==7.4.0
pytest-asyncio==0.22.0
pytest-cov==4.1.0
httpx==0.26.0            # Pour TestClient
faker==22.6.0

# ========== UTILS ==========
python-dateutil==2.8.2
pytz==2024.1

# ========== DEPLOYMENT ==========
gunicorn==21.2.0

# ========== OPTIONAL: PAYMENT GATEWAYS ==========
# stripe==8.2.0          # Décommenter si nécessaire
# cinetpay-sdk==1.0.0    # API paiement mobile CI